applied-ai-system-project/requirements.txt
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
